Added code to add command line arguments #933 	modified:   src/jobs/main_collection_data.py 	modified:   src/jobs/run_main.py
