+++ src/jobs/run_main.py
@@ -23,6 +23,20 @@ This pattern ensures clean separation of concerns and supports scalable, maintai
 """
 
 from jobs.main_collection_data import main
+import argparse
+
+# -------------------- Argument Parser --------------------
+def parse_args():
+    print("parse_args:Parsing command line arguments")
+    parser = argparse.ArgumentParser(description="ETL Process for Collection Data")
+    parser.add_argument("--load_type", type=str, required=True,
+                        help="Type of load operation (e.g., full, incremental)")
+    parser.add_argument("--data_set", type=str, required=True,
+                        help="Name of the dataset to process")
+    args = parser.parse_args()
+    print(f"parse_args:Adding arguments for dataset: {args}")
+    return args
 
 if __name__ == "__main__":
-    main()
+    args = parse_args()
+    main(args)
